fix: move patches in a single folder
